\name{find_res}
\alias{find_res}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Louvain Resolution Detecter}
\description{
finds the resolution that fulfills the is_cluster function
}
\usage{
find_res(tenx, exact = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tenx}{
Seurat object
}
  \item{exact}{
if TRUE, then will take a more precise approach
}
}