\name{gen_matrix_small}
\alias{gen_matrix_small}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
quick matrix of differential genes}
\description{
generates fast matrix of expression for use in clustering}
\usage{
gen_matrix_small(tenx)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tenx}{
Seurat object}
}
\details{}