\name{plot_gen}
\alias{plot_gen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Gene Plot generation
}
\description{
Generates violin and cluster plots for a list of given genes}
\usage{
plot_gen(tenx, list_genes, ex_loc = ".")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tenx}{
input Seurat object
}
  \item{list_genes}{
list of genes to plot in string c() format
}
  \item{ex_loc}{
location of output
}
}
