\name{plot_mito}
\alias{plot_mito}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
plotting mitochondrial gene percentage}
\description{
plots 2 different violin plots for the mitochondria distribution to see quality of data
}
\usage{
plot_mito(cochobj, file_loc)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{cochobj}{
Seurat object
}
  \item{file_loc}{
directory location of violin plot output
}
}